refactor: generatePassword in utils
